Move common/OutputJax.ts to common.js, to be consistent with svg.ts and chtml.ts
