+++ ts/output/common/Wrappers/mroot.ts
@@ -24,7 +24,7 @@
 import {CommonWrapper, CommonWrapperClass, Constructor} from '../Wrapper.js';
 import {CommonWrapperFactory} from '../WrapperFactory.js';
 import {CharOptions, VariantData, DelimiterData, FontData, FontDataClass} from '../FontData.js';
-import {CommonOutputJax} from '../OutputJax.js';
+import {CommonOutputJax} from '../../common.js';
 import {CommonMsqrt, CommonMsqrtClass} from './msqrt.js';
 import {CommonMo} from './mo.js';
 import {BBox} from '../../../util/BBox.js';
